feat(hmr): resolve extensionless dependencies
Note that this has a side-effect of allowing you to hot.accept() an _external node module_. I'm not sure why anyone would ever do this, but it's supported.
